Revert "Create "stdint.h" STL header"
This reverts commit 401ada353954cae803b0137ff0b3d5bfc6744d1f.

The "stdint.h" header was introduced in C99, meaning it's not available under MSVC 6.0 (which released in 1998).

+++ src/gx/GXMisc.c
@@ -2,7 +2,6 @@
 #include "Dolphin/gx.h"
 #include "Dolphin/hw_regs.h"
 #include <stddef.h>
-#include <stdint.h>
 
 static GXDrawSyncCallback TokenCB;
 static GXDrawDoneCallback DrawDoneCB;
@@ -332,7 +331,7 @@ void GXPokeZMode(GXBool compare_enable, GXCompare func, GXBool update_enable)
  */
 void GXPeekARGB(u16 x, u16 y, u32* color)
 {
-	uintptr_t addr = (u32)OSPhysicalToUncached(0x08000000);
+	u32 addr = (u32)OSPhysicalToUncached(0x08000000);
 
 	SET_REG_FIELD(741, addr, 10, 2, x);
 	SET_REG_FIELD(742, addr, 10, 12, y);
@@ -346,7 +345,7 @@ void GXPeekARGB(u16 x, u16 y, u32* color)
  */
 void GXPokeARGB(u16 x, u16 y, u32 color)
 {
-	uintptr_t addr = (u32)OSPhysicalToUncached(0x08000000);
+	u32 addr = (u32)OSPhysicalToUncached(0x08000000);
 
 	SET_REG_FIELD(751, addr, 10, 2, x);
 	SET_REG_FIELD(752, addr, 10, 12, y);
@@ -360,7 +359,7 @@ void GXPokeARGB(u16 x, u16 y, u32 color)
  */
 void GXPeekZ(u16 x, u16 y, u32* z)
 {
-	uintptr_t addr = (u32)OSPhysicalToUncached(0x08000000);
+	u32 addr = (u32)OSPhysicalToUncached(0x08000000);
 
 	SET_REG_FIELD(761, addr, 10, 2, x);
 	SET_REG_FIELD(762, addr, 10, 12, y);
@@ -374,7 +373,7 @@ void GXPeekZ(u16 x, u16 y, u32* z)
  */
 void GXPokeZ(u16 x, u16 y, u32 z)
 {
-	uintptr_t addr = (u32)OSPhysicalToUncached(0x08000000);
+	u32 addr = (u32)OSPhysicalToUncached(0x08000000);
 
 	SET_REG_FIELD(771, addr, 10, 2, x);
 	SET_REG_FIELD(772, addr, 10, 12, y);
